+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com) and this project adheres to [Semantic Versioning](http://semver.org).
 
+## UNRELEASED
+### Changed
+- Plugin setup checks for required options and config values before initializing objects
+
 ## 0.2.1 - 2018-08-07
 ### Added
 - Dist files are now committed to the repo on release